rename brotli -> brotli-codec
