+++ rust/psibase/src/services/mod.rs
@@ -4,6 +4,7 @@ pub mod accounts;
 pub mod auth_delegate;
 pub mod auth_invite;
 pub mod auth_sig;
+pub mod brotli_codec;
 pub mod chainmail;
 pub mod common_api;
 pub mod cpu_limit;
@@ -15,7 +16,6 @@ pub mod invite;
 pub mod nft;
 pub mod packages;
 pub mod producers;
-pub mod psi_brotli;
 pub mod r_events;
 pub mod setcode;
 pub mod sites;
