--- services/user/Brotli/src/lib.rs
@@ -1,1 +1,0 @@
-pub use psi_brotli;
