+++ services/user/BrotliCodec/Cargo.toml
@@ -4,26 +4,19 @@ members = ["service", "plugin"]
 
 [workspace.package]
 version = "0.15.0"
-rust-version = "1.64"
 edition = "2021"
 repository = "https://github.com/gofractally/psibase"
 homepage = "https://psibase.io"
+rust-version = "1.64"
 publish = false
 
 [package]
-name = "brotli_package"
+name = "brotli-codec-pkg"
 description = "Compression/decompression library"
 version.workspace = true
 edition.workspace = true
-
-[package.metadata.psibase]
-package-name = "Brotli"
-services = ["psi-brotli"]
-
-[package.metadata.psibase.dependencies]
-Supervisor = "0.15.0"
-Sites = "0.15.0"
-HttpServer = "0.15.0"
+rust-version.workspace = true
+publish.workspace = true
 
 [profile.release]
 codegen-units = 1
@@ -32,10 +25,19 @@ debug = false
 strip = true
 lto = true
 
+[package.metadata.psibase]
+package-name = "BrotliCodec"
+services = ["brotli-codec"]
+
+[package.metadata.psibase.dependencies]
+Supervisor = "0.15.0"
+Sites = "0.15.0"
+HttpServer = "0.15.0"
+
 [lib]
 crate-type = ["rlib"]
 
 [dependencies]
-psi-brotli = { path = "service", version = "0.15.0" }
+brotli-codec = { path = "service", version = "0.15.0" }
 plugin = { path = "plugin", version = "0.15.0" }
 serde = { version = "1.0", features = ["derive"] }
