+++ services/user/BrotliCodec/service/Cargo.toml
@@ -1,11 +1,11 @@
 [package]
-name = "psi-brotli"
+name = "brotli-codec"
 version.workspace = true
 rust-version.workspace = true
 repository.workspace = true
 homepage.workspace = true
 edition.workspace = true
-publish = false
+publish.workspace = true
 
 [package.metadata.psibase]
 plugin = "plugin"
@@ -14,8 +14,8 @@ plugin = "plugin"
 brotli = { version = "7.0.0" }
 base64 = { version = "0.22" }
 anyhow = { version = "1.0" }
-psibase = { path = "../../../../rust/psibase/" }
+psibase = { path = "../../../../rust/psibase/", version = "0.15.0" }
 serde = { version = "1.0.213", features = ["derive"] }
 
 [dev-dependencies]
-brotli_package = { path = "..", version = "0.15.0" }
+brotli-codec-pkg = { path = "..", version = "0.15.0" }
