+++ services/user/BrotliCodec/src/lib.rs
@@ -1,0 +1,1 @@
+pub use brotli_codec;
